make this work in Firefox and internal chatrooms
So, this is actually kinda handy... But it had two problems: it wouldn't work in the Secret Internal Company Chatrooms™ due to the existence of an extra iframe, and it wouldn't work in Firefox. Also it was a real pain to debug in Firefox. Also it used unsafeWindow, which just makes me uneasy what with the whole unsafe bit. 

So I made a bunch of little changes such that it would work for me. And also screwed with indentation in a few places just for the heck of it. Consider adopting these changes FOR THE GOOD OF HUMANITY, especially the pointless brace-positioning ones.